backend/pyproject.toml
[tool.poetry]
name = "user-management-api"
version = "0.1.0"
description = "FastAPI-based user management service"
authors = ["SDD Team <dev@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
psycopg = { extras = ["binary"], version = "^3.1.19" }
pydantic = "^2.8.2"
pydantic-extra-types = "^2.9.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
anyio = "^4.4.0"
prometheus-fastapi-instrumentator = "^6.1.0"
pydantic-settings = "^2.3.4"
email-validator = "^2.1.0.post1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.7"

[tool.poetry.scripts]
runserver = "app.main:run"
test = "pytest:main"
ci = "pytest:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
